requirements.txt
requests
kafka-python==2.1.4
ansible-core==2.18.3
psycopg2-binary==2.9.10
uvicorn==0.34.0
fastapi==0.115.12
gensim==4.3.3
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
pydantic==2.10.6
scikit_learn==1.3.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.38